requirements.txt
selene>=2.0.b4
pytest~=7.4.3
allure-pytest=2.13.2